requirements.txt
Flask==3.0.0
numpy>=1.21.5,<=1.22
Pillow>=9.2.0,<=10.3.0



